refactor: db connector 인터페이스 이동
