+++ src/main/java/infrastructure/dao/MysqlConnector.java
@@ -1,6 +1,5 @@
 package infrastructure.dao;
 
-import application.persistence.DbConnector;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
